+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.13
 - In both center modes, view can be rotated with held down right mouse button
 - Movement speed in frontal view is now FPS-independent and faster (10 ly per second)
+- Added option to use accurate star positions (taken from a star catalog) instead of reconstructed 
 
 # 0.12
 - Added 20 or so speculative systems